requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
